+++ lib/src/core.dart
@@ -1,4 +1,0 @@
-/// Provides the core functionality.
-library where.core;
-
-part 'core/file_stats.dart';