+++ lib/src/io.dart
@@ -1,5 +1,0 @@
-/// Provides the I/O support.
-library where.io;
-
-export 'io/vm.dart'
-  if (dart.library.js) 'io/node.dart';