+++ test/all.dart
@@ -1,10 +1,10 @@
-import 'core_test.dart' as core_test;
 import 'file_test.dart' as file_test;
 import 'io_test.dart' as io_test;
+import 'file_stats_test.dart' as file_stats_test;
 
 /// Tests all the features of the package.
 void main() {
-  core_test.main();
   file_test.main();
   io_test.main();
+  file_stats_test.main();
 }